+++ src/main/kotlin/ru/mipt/npm/gradle/KScienceExtension.kt
@@ -2,8 +2,6 @@ package ru.mipt.npm.gradle
 
 import org.gradle.api.Project
 import org.gradle.api.plugins.ApplicationPlugin
-import org.gradle.api.publish.maven.plugins.MavenPublishPlugin
-import org.gradle.kotlin.dsl.apply
 import org.gradle.kotlin.dsl.findByType
 import org.jetbrains.kotlin.gradle.dsl.KotlinJsProjectExtension
 import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
@@ -49,7 +47,7 @@ class KScienceExtension(val project: Project) {
     fun useCoroutines(
         version: String = KScienceVersions.coroutinesVersion,
         sourceSet: DependencySourceSet = DependencySourceSet.MAIN,
-        configuration: DependencyConfiguration = DependencyConfiguration.API
+        configuration: DependencyConfiguration = DependencyConfiguration.API,
     ): Unit = project.useCommonDependency(
         "org.jetbrains.kotlinx:kotlinx-coroutines-core:$version",
         dependencySourceSet = sourceSet,
@@ -62,7 +60,7 @@ class KScienceExtension(val project: Project) {
     fun useAtomic(
         version: String = KScienceVersions.atomicVersion,
         sourceSet: DependencySourceSet = DependencySourceSet.MAIN,
-        configuration: DependencyConfiguration = DependencyConfiguration.IMPLEMENTATION
+        configuration: DependencyConfiguration = DependencyConfiguration.IMPLEMENTATION,
     ): Unit = project.run {
         plugins.apply("kotlinx-atomicfu")
         useCommonDependency(
@@ -79,7 +77,7 @@ class KScienceExtension(val project: Project) {
         version: String = KScienceVersions.serializationVersion,
         sourceSet: DependencySourceSet = DependencySourceSet.MAIN,
         configuration: DependencyConfiguration = DependencyConfiguration.API,
-        block: SerializationTargets.() -> Unit = {}
+        block: SerializationTargets.() -> Unit = {},
     ): Unit = project.run {
         plugins.apply("org.jetbrains.kotlin.plugin.serialization")
         val artifactName = if (version.startsWith("0")) {
@@ -95,17 +93,23 @@ class KScienceExtension(val project: Project) {
         SerializationTargets(sourceSet, configuration).apply(block)
     }
 
+    /**
+     * Add platform-specific JavaFX dependencies with given list of [FXModule]s
+     */
     fun useFx(
         vararg modules: FXModule,
         configuration: DependencyConfiguration = DependencyConfiguration.COMPILE_ONLY,
         version: String = "11",
-        platform: FXPlatform = defaultPlatform
+        platform: FXPlatform = defaultPlatform,
     ) = project.useFx(modules.toList(), configuration, version, platform)
 
+    /**
+     * Add dependency on kotlinx-html library
+     */
     fun useHtml(
-        version: String = KScienceVersions.atomicVersion,
+        version: String = KScienceVersions.htmlVersion,
         sourceSet: DependencySourceSet = DependencySourceSet.MAIN,
-        configuration: DependencyConfiguration = DependencyConfiguration.IMPLEMENTATION
+        configuration: DependencyConfiguration = DependencyConfiguration.API,
     ): Unit = project.useCommonDependency(
         "org.jetbrains.kotlinx:kotlinx-html:$version",
         dependencySourceSet = sourceSet,
@@ -115,8 +119,23 @@ class KScienceExtension(val project: Project) {
     /**
      * Use kotlinx-datetime library with default version or [version]
      */
-    fun useDateTime(version: String = KScienceVersions.dateTimeVersion){
-        project.useCommonDependency("org.jetbrains.kotlinx:kotlinx-datetime:$version")
+    fun useDateTime(
+        version: String = KScienceVersions.dateTimeVersion,
+        sourceSet: DependencySourceSet = DependencySourceSet.MAIN,
+        configuration: DependencyConfiguration = DependencyConfiguration.API,
+    ) {
+        project.useCommonDependency(
+            "org.jetbrains.kotlinx:kotlinx-datetime:$version",
+            dependencySourceSet = sourceSet,
+            dependencyConfiguration = configuration
+        )
+    }
+
+    /**
+     * Apply jupyter plugin
+     */
+    fun useJupyter() {
+        project.plugins.apply("org.jetbrains.kotlin.jupyter.api")
     }
 
     /**
@@ -145,10 +164,6 @@ class KScienceExtension(val project: Project) {
                 it.binaries.executable()
             }
         }
-    }
-
-    fun publish() {
-        project.plugins.apply(MavenPublishPlugin::class)
     }
 }