+++ src/main/kotlin/ru/mipt/npm/gradle/KScienceProjectPlugin.kt
@@ -7,6 +7,8 @@ import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.*
 import org.jetbrains.changelog.ChangelogPlugin
+import org.jetbrains.changelog.ChangelogPluginExtension
+import org.jetbrains.dokka.gradle.AbstractDokkaTask
 import org.jetbrains.dokka.gradle.DokkaPlugin
 import org.jetbrains.dokka.gradle.DokkaTask
 import ru.mipt.npm.gradle.internal.*
@@ -15,7 +17,7 @@ import ru.mipt.npm.gradle.internal.*
 class KSciencePublishingExtension(val project: Project) {
     private var initializedFlag = false
 
-    fun configurePublications(vcsUrl: String) {
+    fun vcs(vcsUrl: String) {
         if (!initializedFlag) {
             project.setupPublication(vcsUrl)
             initializedFlag = true
@@ -24,21 +26,37 @@ class KSciencePublishingExtension(val project: Project) {
 
     /**
      * github publishing
+     * @param publish include github packages in release publishing. By default - false
      */
-    fun github(githubProject: String, githubOrg: String = "mipt-npm") {
+    fun github(githubProject: String, githubOrg: String = "mipt-npm", publish: Boolean = false) {
         //automatically initialize vcs using github
         if (!initializedFlag) {
-            configurePublications("https://github.com/$githubOrg/$githubProject")
+            vcs("https://github.com/$githubOrg/$githubProject")
         }
         project.addGithubPublishing(githubOrg, githubProject)
+
+        if (publish) {
+            project.tasks.findByName("publish${project.publicationTarget}ToGithubRepository")?.let {publicationTask->
+                releaseTask?.dependsOn(publicationTask)
+            }
+        }
+    }
+
+    private val releaseTask by lazy {
+        project.tasks.findByName("release")
     }
 
     /**
      *  Space publishing
      */
-    fun space(spaceRepo: String = "https://maven.pkg.jetbrains.space/mipt-npm/p/sci/maven") {
-        require(initializedFlag) { "The publishing is not set up use 'configurePublications' method to do so" }
+    fun space(spaceRepo: String = "https://maven.pkg.jetbrains.space/mipt-npm/p/sci/maven", publish: Boolean = false) {
+        require(initializedFlag) { "The project vcs is not set up use 'vcs' method to do so" }
         project.addSpacePublishing(spaceRepo)
+        if (publish) {
+            project.tasks.findByName("publish${project.publicationTarget}ToSpaceRepository")?.let { publicationTask ->
+                releaseTask?.dependsOn(publicationTask)
+            }
+        }
     }
 
 //    // Bintray publishing
@@ -50,9 +68,15 @@ class KSciencePublishingExtension(val project: Project) {
     /**
      *  Sonatype publishing
      */
-    fun sonatype() {
-        require(initializedFlag) { "The publishing is not set up use 'configurePublications' method to do so" }
+    fun sonatype(publish: Boolean = true) {
+        require(initializedFlag) { "The project vcs is not set up use 'vcs' method to do so" }
         project.addSonatypePublishing()
+        if (publish) {
+            project.tasks.findByName("publish${project.publicationTarget}ToSonatypeRepository")
+                ?.let { publicationTask ->
+                    releaseTask?.dependsOn(publicationTask)
+                }
+        }
     }
 }
 
@@ -63,8 +87,21 @@ class KSciencePublishingExtension(val project: Project) {
 open class KScienceProjectPlugin : Plugin<Project> {
     override fun apply(target: Project): Unit = target.run {
         apply<ChangelogPlugin>()
+
         apply<DokkaPlugin>()
         apply<BinaryCompatibilityValidatorPlugin>()
+
+        afterEvaluate {
+            if (isSnapshot()) {
+                configure<ApiValidationExtension> {
+                    validationDisabled = true
+                }
+            } else {
+                configure<ChangelogPluginExtension> {
+                    version = project.version.toString()
+                }
+            }
+        }
 
         val rootReadmeExtension = KScienceReadmeExtension(this)
         extensions.add("ksciencePublish", KSciencePublishingExtension(this))
@@ -164,7 +201,7 @@ open class KScienceProjectPlugin : Plugin<Project> {
             }
         }
 
-        tasks.withType<DokkaTask> {
+        tasks.withType<AbstractDokkaTask> {
             dependsOn(generateReadme)
         }